# frontend/openapi.json
{
 "components": {
  "schemas": {
...
   "AnomalyDetail": {
    "properties": {
     "advisory": {
      "const": true,
      "default": true,
      "title": "Advisory",
      "type": "boolean"
     },
     "agent": {
      "$ref": "#/components/schemas/AnomalyAgent"
     },
     "context": {
      "$ref": "#/components/schemas/AnomalyContext"
     },
     "features": {
      "items": {
       "$ref": "#/components/schemas/PeerFeature"
      },
      "title": "Features",
      "type": "array"
     },
     "generated_at": {
      "format": "date-time",
      "title": "Generated At",
      "type": "string"
     },
     "history_h": {
      "title": "History H",
      "type": "integer"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "model_version": {
      "title": "Model Version",
      "type": "string"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "peer_count": {
      "title": "Peer Count",
      "type": "integer"
     },
     "peer_group": {
      "title": "Peer Group",
      "type": "string"
     },
     "peer_scores": {
      "$ref": "#/components/schemas/PeerScores"
     },
     "reasons": {
      "items": {
       "$ref": "#/components/schemas/AnomalyReason"
      },
      "title": "Reasons",
      "type": "array"
     },
     "reviewed_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviewed At"
     },
     "reviewed_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviewed By"
     },
     "score": {
      "title": "Score",
      "type": "number"
     },
     "status": {
      "$ref": "#/components/schemas/AnomalyStatus"
     },
     "threshold": {
      "title": "Threshold",
      "type": "number"
     },
     "window_end": {
      "format": "date-time",
      "title": "Window End",
      "type": "string"
     },
     "window_h": {
      "title": "Window H",
      "type": "integer"
     },
     "window_start": {
      "format": "date-time",
      "title": "Window Start",
      "type": "string"
     }
    },
    "required": [
     "id",
     "agent",
     "window_start",
     "window_end",
     "score",
     "threshold",
     "peer_group",
     "reasons",
     "status",
     "reviewed_at",
     "note",
     "model_version",
     "generated_at",
     "peer_count",
     "window_h",
     "history_h",
     "features",
     "peer_scores",
     "context",
     "reviewed_by"
    ],
    "title": "AnomalyDetail",
    "type": "object"
   },
...
   "AnomalyPage": {
    "properties": {
     "advisory": {
      "const": true,
      "default": true,
      "title": "Advisory",
      "type": "boolean"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/AnomalyItem"
      },
      "title": "Items",
      "type": "array"
     },
     "page": {
      "title": "Page",
      "type": "integer"
     },
     "page_size": {
      "title": "Page Size",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "AnomalyPage",
    "type": "object"
   },
...
   "AnomalyReviewIn": {
    "properties": {
     "decision": {
      "enum": [
       "confirmed",
       "dismissed"
      ],
      "title": "Decision",
      "type": "string"
     },
     "note": {
      "maxLength": 500,
      "minLength": 1,
      "title": "Note",
      "type": "string"
     }
    },
    "required": [
     "decision",
     "note"
    ],
    "title": "AnomalyReviewIn",
    "type": "object"
   },
   "AnomalyStatus": {
    "enum": [
     "open",
     "confirmed",
     "dismissed"
    ],
    "title": "AnomalyStatus",
    "type": "string"
   },
   "AssumptionsDoc": {
    "properties": {
     "markdown": {
      "title": "Markdown",
      "type": "string"
     },
     "title": {
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "title",
     "markdown"
    ],
    "title": "AssumptionsDoc",
    "type": "object"
   },
...
   "DataCount": {
    "properties": {
     "rows": {
      "title": "Rows",
      "type": "integer"
     },
     "table": {
      "title": "Table",
      "type": "string"
     }
    },
    "required": [
     "table",
     "rows"
    ],
    "title": "DataCount",
    "type": "object"
   },
...
   "DataPeriod": {
    "description": "Synthetic data window; the last 14 days are held out from training.",
    "properties": {
     "end": {
      "format": "date-time",
      "title": "End",
      "type": "string"
     },
     "holdout_start": {
      "format": "date-time",
      "title": "Holdout Start",
      "type": "string"
     },
     "sim_now": {
      "format": "date-time",
      "title": "Sim Now",
      "type": "string"
     },
     "start": {
      "format": "date-time",
      "title": "Start",
      "type": "string"
     }
    },
    "required": [
     "start",
     "end",
     "holdout_start",
     "sim_now"
    ],
    "title": "DataPeriod",
    "type": "object"
   },
   "DataSummary": {
    "properties": {
     "configured_seed": {
      "title": "Configured Seed",
      "type": "integer"
     },
     "counts": {
      "items": {
       "$ref": "#/components/schemas/DataCount"
      },
      "title": "Counts",
      "type": "array"
     },
     "data_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Data Version"
     },
     "generated_at": {
      "format": "date-time",
      "title": "Generated At",
      "type": "string"
     },
     "labelled_anomalous_agents": {
      "title": "Labelled Anomalous Agents",
      "type": "integer"
     },
     "period": {
      "$ref": "#/components/schemas/DataPeriod"
     },
     "seed": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seed"
     }
    },
    "required": [
     "seed",
     "configured_seed",
     "data_version",
     "period",
     "counts",
     "labelled_anomalous_agents",
     "generated_at"
    ],
    "title": "DataSummary",
    "type": "object"
   },
...
   "DriftFloat": {
    "properties": {
     "buckets": {
      "items": {
       "$ref": "#/components/schemas/DriftBucket"
      },
      "title": "Buckets",
      "type": "array"
     },
     "by_horizon": {
      "items": {
       "$ref": "#/components/schemas/DriftPoint"
      },
      "title": "By Horizon",
      "type": "array"
     },
     "float_type": {
      "$ref": "#/components/schemas/FloatType"
     },
     "hours_compared": {
      "title": "Hours Compared",
      "type": "integer"
     },
     "status": {
      "enum": [
       "stable",
       "watch",
       "drift",
       "no_data"
      ],
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "float_type",
     "buckets",
     "by_horizon",
     "hours_compared",
     "status"
    ],
    "title": "DriftFloat",
    "type": "object"
   },
   "DriftPoint": {
    "properties": {
     "horizon_h": {
      "title": "Horizon H",
      "type": "integer"
     },
     "mae": {
      "title": "Mae",
      "type": "number"
     }
    },
    "required": [
     "horizon_h",
     "mae"
    ],
    "title": "DriftPoint",
    "type": "object"
   },
   "DriftReport": {
    "properties": {
     "drift_ratio": {
      "title": "Drift Ratio",
      "type": "number"
     },
     "floats": {
      "items": {
       "$ref": "#/components/schemas/DriftFloat"
      },
      "title": "Floats",
      "type": "array"
     },
     "generated_at": {
      "format": "date-time",
      "title": "Generated At",
      "type": "string"
     },
     "model_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model Version"
     },
     "origin": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Origin"
     },
     "status": {
      "enum": [
       "stable",
       "watch",
       "drift",
       "no_data"
      ],
      "title": "Status",
      "type": "string"
     },
     "watch_ratio": {
      "title": "Watch Ratio",
      "type": "number"
     }
    },
    "required": [
     "model_version",
     "origin",
     "watch_ratio",
     "drift_ratio",
     "floats",
     "status",
     "generated_at"
    ],
    "title": "DriftReport",
    "type": "object"
   },
...
   "GeneratedBy": {
    "enum": [
     "llm",
     "template",
     "replay"
    ],
    "title": "GeneratedBy",
    "type": "string"
   },
...
   "GuardResult": {
    "enum": [
     "pass",
     "numbers_fail",
     "injection",
     "schema_fail",
     "timeout",
     "error"
    ],
    "title": "GuardResult",
    "type": "string"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "JobIn": {
    "additionalProperties": false,
    "properties": {
     "kind": {
      "enum": [
       "generate_data",
       "retrain_forecast",
       "retrain_anomaly"
      ],
      "title": "Kind",
      "type": "string"
     }
    },
    "required": [
     "kind"
    ],
    "title": "JobIn",
    "type": "object"
   },
   "JobList": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/JobOut"
      },
      "title": "Items",
      "type": "array"
     },
     "running": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/JobOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "required": [
     "items",
     "running"
    ],
    "title": "JobList",
    "type": "object"
   },
   "JobOut": {
    "properties": {
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "finished_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "kind": {
      "enum": [
       "generate_data",
       "retrain_forecast",
       "retrain_anomaly"
      ],
      "title": "Kind",
      "type": "string"
     },
     "progress": {
      "title": "Progress",
      "type": "integer"
     },
     "result": {
      "items": {
       "$ref": "#/components/schemas/JobResultItem"
      },
      "title": "Result",
      "type": "array"
     },
     "started_by_email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started By Email"
     },
     "status": {
      "enum": [
       "queued",
       "running",
       "succeeded",
       "failed"
      ],
      "title": "Status",
      "type": "string"
     },
     "step": {
      "title": "Step",
      "type": "string"
     },
     "updated_at": {
      "format": "date-time",
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "id",
     "kind",
     "status",
     "progress",
     "step",
     "error",
     "result",
     "started_by_email",
     "created_at",
     "updated_at",
     "finished_at"
    ],
    "title": "JobOut",
    "type": "object"
   },
   "JobResultItem": {
    "description": "One labelled outcome value (row counts, model version, metric deltas).",
    "properties": {
     "key": {
      "title": "Key",
      "type": "string"
     },
     "value": {
      "title": "Value",
      "type": "string"
     }
    },
    "required": [
     "key",
     "value"
    ],
    "title": "JobResultItem",
    "type": "object"
   },
...
   "LlmIntent": {
    "enum": [
     "copilot",
     "narrate",
     "agent_briefing",
     "distributor_briefing",
     "anomaly_narrative"
    ],
    "title": "LlmIntent",
    "type": "string"
   },
...
   "LlmLogPage": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/LlmLogItem"
      },
      "title": "Items",
      "type": "array"
     },
     "page": {
      "title": "Page",
      "type": "integer"
     },
     "page_size": {
      "title": "Page Size",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "LlmLogPage",
    "type": "object"
   },
...
   "LlmUsage": {
    "properties": {
     "avg_latency_ms": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avg Latency Ms"
     },
     "cache_hit_rate": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cache Hit Rate"
     },
     "calls_today": {
      "title": "Calls Today",
      "type": "integer"
     },
     "cap_used": {
      "title": "Cap Used",
      "type": "number"
     },
     "daily_cap": {
      "title": "Daily Cap",
      "type": "integer"
     },
     "days": {
      "items": {
       "$ref": "#/components/schemas/LlmUsageDay"
      },
      "title": "Days",
      "type": "array"
     },
     "generated_at": {
      "format": "date-time",
      "title": "Generated At",
      "type": "string"
     },
     "guard_failures": {
      "title": "Guard Failures",
      "type": "integer"
     },
     "p95_latency_ms": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "P95 Latency Ms"
     },
     "total_calls": {
      "title": "Total Calls",
      "type": "integer"
     }
    },
    "required": [
     "days",
     "calls_today",
     "daily_cap",
     "cap_used",
     "total_calls",
     "cache_hit_rate",
     "avg_latency_ms",
     "p95_latency_ms",
     "guard_failures",
     "generated_at"
    ],
    "title": "LlmUsage",
    "type": "object"
   },
   "LlmUsageDay": {
    "properties": {
     "cache_hits": {
      "title": "Cache Hits",
      "type": "integer"
     },
     "calls": {
      "title": "Calls",
      "type": "integer"
     },
     "completion_tokens": {
      "title": "Completion Tokens",
      "type": "integer"
     },
     "day": {
      "format": "date",
      "title": "Day",
      "type": "string"
     },
     "guard_failures": {
      "title": "Guard Failures",
      "type": "integer"
     },
     "live_calls": {
      "title": "Live Calls",
      "type": "integer"
     },
     "prompt_tokens": {
      "title": "Prompt Tokens",
      "type": "integer"
     },
     "replay": {
      "title": "Replay",
      "type": "integer"
     },
     "template": {
      "title": "Template",
      "type": "integer"
     }
    },
    "required": [
     "day",
     "calls",
     "live_calls",
     "cache_hits",
     "replay",
     "template",
     "guard_failures",
     "prompt_tokens",
     "completion_tokens"
    ],
    "title": "LlmUsageDay",
    "type": "object"
   },
...
   "ModelRegistry": {
    "properties": {
     "generated_at": {
      "format": "date-time",
      "title": "Generated At",
      "type": "string"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/ModelVersionItem"
      },
      "title": "Items",
      "type": "array"
     }
    },
    "required": [
     "items",
     "generated_at"
    ],
    "title": "ModelRegistry",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "title": "AgentPulse AI",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/admin/data": {
   "get": {
    "description": "Seed, data version, period (holdout, SIM_NOW) and row counts of the synthetic dataset.",
    "operationId": "get_data_api_v1_admin_data_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DataSummary"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Data",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/admin/data/assumptions": {
   "get": {
    "description": "docs/SYNTHETIC_ASSUMPTIONS.md as Markdown text.",
    "operationId": "get_assumptions_api_v1_admin_data_assumptions_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AssumptionsDoc"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Assumptions",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/admin/drift": {
   "get": {
    "description": "Forecast-error drift: cached forecast vs logged demand, against the holdout MAE.",
    "operationId": "get_drift_api_v1_admin_drift_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DriftReport"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Drift",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/admin/jobs": {
   "get": {
    "description": "The 20 most recent jobs and the one running, if any.",
    "operationId": "list_jobs_api_v1_admin_jobs_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobList"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Jobs",
    "tags": [
     "admin"
    ]
   },
   "post": {
    "description": "Start generate_data, retrain_forecast or retrain_anomaly in the background; poll\nGET /admin/jobs/{id} for progress. One job at a time (409 job_running). Retrained models are\nrecorded inactive; serving keeps the committed model.",
    "operationId": "start_job_api_v1_admin_jobs_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/JobIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Start Job",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/admin/jobs/{job_id}": {
   "get": {
    "operationId": "get_job_api_v1_admin_jobs__job_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "job_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Job Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Job",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/admin/llm/logs": {
   "get": {
    "description": "llm_call_log newest first: tokens, latency, generated_by, cache hit, guard result.",
    "operationId": "llm_logs_api_v1_admin_llm_logs_get",
    "parameters": [
     {
      "in": "query",
      "name": "intent",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/LlmIntent"
        },
        {
         "type": "null"
        }
       ],
       "title": "Intent"
      }
     },
     {
      "in": "query",
      "name": "generated_by",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/GeneratedBy"
        },
        {
         "type": "null"
        }
       ],
       "title": "Generated By"
      }
     },
     {
      "in": "query",
      "name": "guard_result",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/GuardResult"
        },
        {
         "type": "null"
        }
       ],
       "title": "Guard Result"
      }
     },
     {
      "in": "query",
      "name": "cache_hit",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cache Hit"
      }
     },
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "maximum": 10000,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 50,
       "maximum": 200,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LlmLogPage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Llm Logs",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/admin/llm/usage": {
   "get": {
    "description": "Calls per UTC day by outcome, tokens, latency, cache hit rate and today's cap usage.",
    "operationId": "llm_usage_api_v1_admin_llm_usage_get",
    "parameters": [
     {
      "in": "query",
      "name": "days",
      "required": false,
      "schema": {
       "default": 7,
       "maximum": 30,
       "minimum": 1,
       "title": "Days",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LlmUsage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Llm Usage",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/admin/models": {
   "get": {
    "description": "Every registered model version (active first) with its stored holdout metrics.",
    "operationId": "get_models_api_v1_admin_models_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ModelRegistry"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Models",
    "tags": [
     "admin"
    ]
   }
  },
...
  "/api/v1/anomalies": {
   "get": {
    "description": "Isolation Forest flags in scope (distributor: own agents; admin: all), open first.",
    "operationId": "list_anomalies_api_v1_anomalies_get",
    "parameters": [
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/AnomalyStatus"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "maximum": 10000,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 20,
       "maximum": 100,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AnomalyPage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Anomalies",
    "tags": [
     "anomalies"
    ]
   }
  },
  "/api/v1/anomalies/{anomaly_id}": {
   "get": {
    "description": "One flag with its peer-group distribution per feature and the top reasons.",
    "operationId": "get_anomaly_api_v1_anomalies__anomaly_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "anomaly_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Anomaly Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AnomalyDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Anomaly",
    "tags": [
     "anomalies"
    ]
   }
  },
...
  "/api/v1/anomalies/{anomaly_id}/review": {
   "post": {
    "description": "Human review: confirmed or dismissed with a note (audit_log). Nothing else happens.",
    "operationId": "review_anomaly_api_v1_anomalies__anomaly_id__review_post",
    "parameters": [
     {
      "in": "path",
      "name": "anomaly_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Anomaly Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AnomalyReviewIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AnomalyDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Review Anomaly",
    "tags": [
     "anomalies"
    ]
   }
  },
...
 }
}